+++ SmokeOrFire/SmokeOrFire/Enums/Rank.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2016 Justin Lawrence Hester. All rights reserved.
 //
 
-enum Rank: UInt8 {
+enum Rank: Int {
     case ACE = 1, TWO, THREE, FOUR, FIVE, SIX, SEVEN, EIGHT, NINE, TEN, JACK, QUEEN, KING, __EXHAUST
 
     static let allValues = [ACE, TWO, THREE, FOUR, FIVE, SIX, SEVEN, EIGHT, NINE, TEN, JACK, QUEEN, KING]
